+++ backend-rust/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- Remove locked CCD metrics
+
 ## [0.1.24] - 2025-02-12
 
 ### Fixed